Store state information for Slack buttons instead of passing it directly
Currently data related to Slack actions is stringified then passed into the Slack API.

The Slack API limits this data to 75 characters, which could cause issues with long names

This commit uses assigns a unique value to each object that would have been stringified and stores it locally in a map.
It is then retrieved using the same ID when Slack calls back to the app.